Deprecate `"g3w-ol/utils"` folder
- deprecate notice for "g3w-ol/utils" folder (will be removed after v3.4)
- deprecate notice for "g3w-ol/utils.js" file (will be removed after v3.4)
- move whole utility functions within "core/utils/ol"
