+++ src/app/g3w-ol/controls/querycontrol.js
@@ -1,4 +1,4 @@
-const utils = require('../utils');
+const utils = require('core/utils/ol');
 const InteractionControl = require('./interactioncontrol');
 const PickCoordinatesInteraction = require('../interactions/pickcoordinatesinteraction');
