+++ src/app/g3w-ol/g3w.ol.js
@@ -1,14 +1,8 @@
-const utils = require('./utils');
+const utils = require('core/utils/ol');
 const maphelpers = require('./map/maphelpers');
 
-(function (name, root, factory) {
-  if (typeof define === 'function' && define.amd) define(factory);
-  else if (typeof exports === 'object') module.exports = factory();
-  else root[name] = factory();
-})('g3wol3', this, function () {
-  'use strict';
-  const helpers = utils.merge({}, maphelpers);
-  return {
-    helpers: helpers
-  }
-});
+const helpers = utils.merge({}, maphelpers);
+
+module.exports = {
+  helpers: helpers
+};
