+++ src/app/gui/print/printservice.js
@@ -5,7 +5,7 @@ const GUI = require('gui/gui');
 const G3WObject = require('core/g3wobject');
 const ProjectsRegistry = require('core/project/projectsregistry');
 const PrintService = require('core/print/printservice');
-const {getScaleFromResolution, getResolutionFromScale, getMetersFromDegrees} = require('g3w-ol/utils/utils');
+const {getScaleFromResolution, getResolutionFromScale, getMetersFromDegrees} = require('core/utils/ol');
 const printConfig = require('./printconfig');
 const PrintPage = require('./vue/printpage');
 const scale = printConfig.scale;
